BugFix
The first bug I fixed was related to the header of the web page. There was a div element with the class ml-10 space-x-8 that was causing the HTML to break when the page was viewed on smaller screens.

The second bug I fixed was related to the color of a logo on the web page. The original logo was black, now its changed to white.